+++ admin/users.php
@@ -16,23 +16,23 @@ $rs_user = new User;
 	switch($action) {
 		case 'create':
 			// Create a new user
-			$rs_user->createEntry();
+			userHasPrivilege($_SESSION['role'], 'can_create_users') ? $rs_user->createEntry() : redirect('users.php');
 			break;
 		case 'edit':
 			// Edit an existing user
-			$rs_user->editEntry($id);
+			userHasPrivilege($_SESSION['role'], 'can_edit_users') ? $rs_user->editEntry($id) : redirect('users.php');
 			break;
 		case 'delete':
 			// Delete an existing user
-			$rs_user->deleteEntry($id);
+			userHasPrivilege($_SESSION['role'], 'can_delete_users') ? $rs_user->deleteEntry($id) : redirect('users.php');
 			break;
 		case 'reset_password':
 			// Reset a user's password
-			$rs_user->resetPassword($id);
+			userHasPrivilege($_SESSION['role'], 'can_edit_users') ? $rs_user->resetPassword($id) : redirect('users.php');
 			break;
 		default:
 			// List all users
-			$rs_user->listEntries();
+			userHasPrivilege($_SESSION['role'], 'can_view_users') ? $rs_user->listEntries() : redirect('index.php');
 	}
 	?>
 </div>