+++ admin/widgets.php
@@ -16,19 +16,19 @@ $rs_widget = new Widget;
 	switch($action) {
 		case 'create':
 			// Create a new widget
-			$rs_widget->createEntry();
+			userHasPrivilege($_SESSION['role'], 'can_create_widgets') ? $rs_widget->createEntry() : redirect('widgets.php');
 			break;
 		case 'edit':
 			// Edit an existing widget
-			$rs_widget->editEntry($id);
+			userHasPrivilege($_SESSION['role'], 'can_edit_widgets') ? $rs_widget->editEntry($id) : redirect('widgets.php');
 			break;
 		case 'delete':
 			// Delete an existing widget
-			$rs_widget->deleteEntry($id);
+			userHasPrivilege($_SESSION['role'], 'can_delete_widgets') ? $rs_widget->deleteEntry($id) : redirect('widgets.php');
 			break;
 		default:
 			// List all widgets
-			$rs_widget->listEntries();
+			userHasPrivilege($_SESSION['role'], 'can_view_widgets') ? $rs_widget->listEntries() : redirect('index.php');
 	}
 	?>
 </div>